provide an option for over provisioning of services
Users generally don't use 100% of their quota.  This
is an initial implementation of over provisioning
support.  For now, it is binary.  In the future,
we might add a feature that sets the max allowable
over provisioning amount, likely expressed
as a % of the committed capacity.

(cherry pick of anfernee/vcap-services b731d2cdac78a78d10ef
with minor edits, including disallowing over provisioning
by default)