Update rabbit_node.rb
According to the documentation this field should be URI instead of URL. I created an issue for that in service-contrib-release project. This is a suggestion due to the fact that the other services as well as the central run.pivotal.io instance returns URI and not URL for a provisioned and configured service binding.